Struct xml_tree::ParserError

pub struct ParserError {
    // some fields omitted
}

XML parsing error.

Consists of a row and column reference and a message.

Methods

impl Error

fn new<O>(o: &O, msg: String) -> Error where O: HasPosition

Creates a new error using position information from the provided HasPosition object and a message.

fn new_full(row: usize, col: usize, msg: String) -> Error

Creates a new error using provided position information and a message.

fn msg(&'a self) -> &'a str

Returns a reference to a message which is contained inside this error.